+++ lib/axis_layout.dart
@@ -5,4 +5,3 @@ export 'src/main_alignment.dart';
 export 'src/cross_alignment.dart';
 export 'src/parent_data.dart';
 export 'src/axis_layout_child.dart';
-export 'src/shrink_order.dart';