+++ lib/src/parent_data.dart
@@ -5,4 +5,5 @@ class AxisLayoutParentData extends ContainerBoxParentData<RenderBox> {
   double fill = 0;
   double shrink = 0;
   double fit = 0;
+  int shrinkOrder = 0;
 }