+++ lib/src/shrink_order.dart
@@ -1,1 +1,0 @@
-enum ShrinkOrder { forward, reverse, simultaneous }