+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Horizontal layout
+* Shrink weight
+* Shrink order
+
 ## 0.0.1
 
 * Package creation